moved requests.py into `utilities/`
