+++ libs/langchain/langchain/agents/agent_toolkits/openapi/planner.py
@@ -37,7 +37,7 @@ from langchain.schema import BasePromptTemplate
 from langchain.schema.language_model import BaseLanguageModel
 from langchain.tools.base import BaseTool
 from langchain.tools.requests.tool import BaseRequestsTool
-from langchain.utils.requests import RequestsWrapper
+from langchain.utilities.requests import RequestsWrapper
 
 #
 # Requests tools with LLM-instructed extraction of truncated responses.
