+++ libs/langchain/langchain/agents/agent_toolkits/openapi/toolkit.py
@@ -19,7 +19,7 @@ from langchain.tools.requests.tool import (
     RequestsPostTool,
     RequestsPutTool,
 )
-from langchain.utils.requests import TextRequestsWrapper
+from langchain.utilities.requests import TextRequestsWrapper
 
 
 class RequestsToolkit(BaseToolkit):
