+++ libs/langchain/langchain/chains/llm_requests.py
@@ -8,7 +8,7 @@ from pydantic import Extra, Field, root_validator
 from langchain.callbacks.manager import CallbackManagerForChainRun
 from langchain.chains import LLMChain
 from langchain.chains.base import Chain
-from langchain.utils.requests import TextRequestsWrapper
+from langchain.utilities.requests import TextRequestsWrapper
 
 DEFAULT_HEADERS = {
     "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/87.0.4280.88 Safari/537.36"  # noqa: E501
