+++ libs/langchain/langchain/requests.py
@@ -1,5 +1,5 @@
 """DEPRECATED: Kept for backwards compatibility."""
-from langchain.utils import Requests, RequestsWrapper, TextRequestsWrapper
+from langchain.utilities import Requests, RequestsWrapper, TextRequestsWrapper
 
 __all__ = [
     "Requests",
